dev-env job changes

+++ src/main/java/com/facilio/tasker/FacilioTimer.java
@@ -85,7 +85,7 @@ public class FacilioTimer {
 		if (AccountUtil.getCurrentAccount() != null) {
 			jc.setTimezone(AccountUtil.getCurrentAccount().getTimeZone());
 		}
-		if(!FacilioProperties.isProduction()){
+		if(!FacilioProperties.isProduction() && !FacilioProperties.isDevelopment()){
 			FacilioService.runAsService(FacilioConstants.Services.TEMP_JOBS,() -> JobStore.addJob(jc));
 		}else {
 			FacilioService.runAsService(FacilioConstants.Services.JOB_SERVICE,() -> JobStore.addJob(jc));
@@ -122,7 +122,7 @@ public class FacilioTimer {
 		if (AccountUtil.getCurrentAccount() != null) {
 			jc.setTimezone(AccountUtil.getCurrentAccount().getTimeZone());
 		}
-		if(!FacilioProperties.isProduction()){
+		if(!FacilioProperties.isProduction() && !FacilioProperties.isDevelopment()){
 			FacilioService.runAsService(FacilioConstants.Services.TEMP_JOBS,() -> JobStore.addJob(jc));
 		}else {
 			FacilioService.runAsService(FacilioConstants.Services.JOB_SERVICE,() -> JobStore.addJob(jc));
@@ -131,7 +131,7 @@ public class FacilioTimer {
 	
 	public static void deleteJob(long jobId, String jobName) throws Exception {
 		long orgId = getCurrentOrgId();
-		if(!FacilioProperties.isProduction() ){
+		if(!FacilioProperties.isProduction() && !FacilioProperties.isDevelopment() ){
 			FacilioService.runAsService(FacilioConstants.Services.TEMP_JOBS,() -> JobStore.deleteJob(orgId, jobId, jobName));
 		}else {
 			FacilioService.runAsService(FacilioConstants.Services.JOB_SERVICE,() -> JobStore.deleteJob(orgId, jobId, jobName));
@@ -140,7 +140,7 @@ public class FacilioTimer {
 
 	public static void deleteJobs ( List<Long> jobIds,String jobName ) throws Exception {
 		long orgId = getCurrentOrgId();
-		if(!FacilioProperties.isProduction() ) {
+		if(!FacilioProperties.isProduction() && !FacilioProperties.isDevelopment()) {
 			FacilioService.runAsService(FacilioConstants.Services.TEMP_JOBS,() -> JobStore.deleteJobs(orgId,jobIds,jobName));
 		}else {
 			FacilioService.runAsService(FacilioConstants.Services.JOB_SERVICE,() -> JobStore.deleteJobs(orgId,jobIds,jobName));
@@ -149,7 +149,7 @@ public class FacilioTimer {
 	
 	public static JobContext getJob(long jobId, String jobName) throws Exception {
 		long orgId = getCurrentOrgId();
-		if(!FacilioProperties.isProduction()){
+		if(!FacilioProperties.isProduction() && !FacilioProperties.isDevelopment()){
 			return FacilioService.runAsServiceWihReturn(FacilioConstants.Services.TEMP_JOBS,() -> JobStore.getJob(orgId, jobId, jobName));
 		}else {
 			return FacilioService.runAsServiceWihReturn(FacilioConstants.Services.JOB_SERVICE,() -> JobStore.getJob(orgId, jobId, jobName));
@@ -158,7 +158,7 @@ public class FacilioTimer {
 	
 	public static List<JobContext> getJobs(List<Long> jobIds, String jobName) throws Exception {
 		long orgId = getCurrentOrgId();
-		if(!FacilioProperties.isProduction()){
+		if(!FacilioProperties.isProduction() && !FacilioProperties.isDevelopment()){
 			return FacilioService.runAsServiceWihReturn(FacilioConstants.Services.TEMP_JOBS,() -> JobStore.getJobs(orgId, jobIds, jobName));
 		}else {
 			return FacilioService.runAsServiceWihReturn(FacilioConstants.Services.JOB_SERVICE,() -> JobStore.getJobs(orgId, jobIds, jobName));
@@ -167,7 +167,7 @@ public class FacilioTimer {
 
 	public static int activateJob ( long jobId,String jobName ) throws Exception {
 		long orgId = getCurrentOrgId();
-		if(!FacilioProperties.isProduction()) {
+		if(!FacilioProperties.isProduction() && !FacilioProperties.isDevelopment()) {
 			return FacilioService.runAsServiceWihReturn(FacilioConstants.Services.TEMP_JOBS,() -> JobStore.setStatusForJob(orgId,jobId,jobName,true));
 		}else {
 			return FacilioService.runAsServiceWihReturn(FacilioConstants.Services.JOB_SERVICE,() -> JobStore.setStatusForJob(orgId,jobId,jobName,true));
@@ -176,7 +176,7 @@ public class FacilioTimer {
 
 	public static int inActivateJob ( long jobId,String jobName ) throws Exception {
 		long orgId = getCurrentOrgId();
-		if(!FacilioProperties.isProduction()) {
+		if(!FacilioProperties.isProduction() && !FacilioProperties.isDevelopment()) {
 			return FacilioService.runAsServiceWihReturn(FacilioConstants.Services.TEMP_JOBS,() -> JobStore.setStatusForJob(orgId,jobId,jobName,false));
 		}else {
 			return FacilioService.runAsServiceWihReturn(FacilioConstants.Services.JOB_SERVICE,() -> JobStore.setStatusForJob(orgId,jobId,jobName,false));
